# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HTML Link Parser and Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ParseUrlInput {
  /** The URL to fetch and parse */
  url: string;
}

export interface ParseFileInput {
  /**
     * Raw HTML content string to parse (for .html/.htm files)
     * @nullable
     */
  html_content?: string | null;
  /**
     * Base64-encoded file content (for binary formats like .webarchive)
     * @nullable
     */
  file_content_b64?: string | null;
  /**
     * Optional filename for display purposes and format detection
     * @nullable
     */
  filename?: string | null;
  /**
     * Optional base URL for resolving relative links
     * @nullable
     */
  base_url?: string | null;
}

/**
 * Link category
 */
export type ParsedLinkType = typeof ParsedLinkType[keyof typeof ParsedLinkType];


export const ParsedLinkType = {
  internal: 'internal',
  external: 'external',
  anchor: 'anchor',
  special: 'special',
} as const;

export interface ParsedLink {
  /** Sequential link ID */
  id: number;
  /** Visible link text */
  text: string;
  /** Original href attribute */
  href: string;
  /** Absolute resolved URL */
  resolved_href: string;
  /** Link category */
  type: ParsedLinkType;
  /**
     * Semantic section the link belongs to (header, nav, main, footer, etc.)
     * @nullable
     */
  section: string | null;
  /**
     * Nearest preceding heading in the same section
     * @nullable
     */
  heading: string | null;
  /** DOM order position of the link */
  position: number;
}

export interface LinkMetrics {
  total: number;
  internal: number;
  external: number;
  anchor: number;
  special: number;
}

export interface HeadingGroup {
  /**
     * Heading text (null if no heading)
     * @nullable
     */
  heading: string | null;
  links: ParsedLink[];
}

export interface GroupedSection {
  /** Name of the semantic section */
  section: string;
  headings: HeadingGroup[];
}

export interface ParseResult {
  /** The source identifier (URL or filename) */
  source: string;
  /** Base URL used for resolving relative links */
  base_url: string;
  links: ParsedLink[];
  metrics: LinkMetrics;
  grouped: GroupedSection[];
}

export interface ErrorResponse {
  error: string;
}