# lib/api-zod/src/generated/types/parseFileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HTML Link Parser and Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface ParseFileInput {
  /**
     * Raw HTML content string to parse (for .html/.htm files)
     * @nullable
     */
  html_content?: string | null;
  /**
     * Base64-encoded file content (for binary formats like .webarchive)
     * @nullable
     */
  file_content_b64?: string | null;
  /**
     * Optional filename for display purposes and format detection
     * @nullable
     */
  filename?: string | null;
  /**
     * Optional base URL for resolving relative links
     * @nullable
     */
  base_url?: string | null;
}